Extract random table selection logic to its own helper
Also modify its behavior to throw an ActionException when it can't
select a failed table, to improve the statistics. With this change if an
action is skipped because a failed lookup, it will show up in the
failure section of statistics instead of success.
