+++ libstormweaver/include/action/action.hpp
@@ -20,4 +20,14 @@ public:
                        sql_variant::LoggedSQL *connection) const = 0;
 };
 
+class ActionException : public std::exception {
+public:
+  ActionException(std::string const &message) : message(message) {}
+
+  const char *what() const noexcept override { return message.c_str(); }
+
+private:
+  std::string message;
+};
+
 } // namespace action
