+++ libstormweaver/include/action/helper.hpp
@@ -1,0 +1,12 @@
+
+#pragma once
+
+#include "action/action.hpp"
+#include "metadata.hpp"
+
+namespace action {
+
+metadata::table_cptr find_random_table(metadata::Metadata const &metaCtx,
+                                       ps_random &rand);
+
+}
